% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr_process_subfuncs.R
\name{compute_dr_ssanom}
\alias{compute_dr_ssanom}
\title{Single Site Anomaly Detection method for DR}
\usage{
compute_dr_ssanom(cohort, dr_ptlv_rslt, n_sd = 2)
}
\arguments{
\item{cohort}{table of cohort members with at least \code{site}, \code{person_id}, \code{start_date}, and \code{end_date}}

\item{dr_ptlv_rslt}{patient level results output by \code{dr_process}}

\item{n_sd}{numeric indicating the number of standard deviations away from the mean that will indicate an outlier (defaults to 2)}
}
\description{
Single Site Anomaly Detection method for DR
}
\keyword{internal}
