% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr_output_subfuncs.R
\name{dr_ms_anom_la}
\alias{dr_ms_anom_la}
\title{\emph{Multi Site, Anomaly Detection, Longitudinal}}
\usage{
dr_ms_anom_la(
  process_output,
  filter_definition = NULL,
  text_wrapping_char = 60L,
  large_n = FALSE,
  large_n_sites = NULL
)
}
\arguments{
\item{process_output}{tabular output from \code{dr_process}}

\item{filter_definition}{the name(s) of duplicate_definitions that should be included on the plot}

\item{text_wrapping_char}{an integer indicating the length limit for text wrapping on axis text}

\item{large_n}{a boolean indicating whether the large N visualization, intended for a high
volume of sites, should be used; defaults to FALSE}

\item{large_n_sites}{a vector of site names that can optionally generate a filtered visualization}
}
\value{
three graphs:
\enumerate{
\item line graph that shows the smoothed proportion of
unmapped concepts across time computation with the Euclidean distance associated with each line
\item line graph that shows the raw proportion of
unmapped concepts across time computation with the Euclidean distance associated with each line
\item a bar graph with the Euclidean distance value for each site, with the average
proportion as the fill
}
}
\description{
\emph{Multi Site, Anomaly Detection, Longitudinal}
}
\keyword{internal}
