% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr_output_subfuncs.R
\name{dr_ms_exp_la}
\alias{dr_ms_exp_la}
\title{\emph{Multi Site, Exploratory, Longitudinal}}
\usage{
dr_ms_exp_la(
  process_output,
  output_col,
  filter_definition = NULL,
  text_wrapping_char = 60L,
  facet = NULL,
  large_n = FALSE,
  large_n_sites = NULL
)
}
\arguments{
\item{process_output}{tabular output from \code{dr_process}}

\item{output_col}{name of the column from process_output to be used in the analysis.
can be any of the proportion or either of the median_site_* columns}

\item{filter_definition}{the name(s) of duplicate_definitions that should be included on the plot}

\item{text_wrapping_char}{an integer indicating the length limit for text wrapping on axis text}

\item{facet}{fields by which the graph should be facetted}

\item{large_n}{a boolean indicating whether the large N visualization, intended for a high
volume of sites, should be used; defaults to FALSE}

\item{large_n_sites}{a vector of site names that can optionally generate a filtered visualization}
}
\value{
a line graph showing the proportion or median per patient of
unmapped concepts per site & duplicate_definition across the time series
}
\description{
\emph{Multi Site, Exploratory, Longitudinal}
}
\keyword{internal}
