% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr_output.R
\name{dr_output}
\alias{dr_output}
\title{Duplicate Records -- Output Generation}
\usage{
dr_output(
  process_output,
  output_col,
  filter_definition = NULL,
  text_wrapping_char = 60L,
  large_n = FALSE,
  large_n_sites = NULL
)
}
\arguments{
\item{process_output}{\emph{tabular input} || \strong{required}

The tabular output produced by \code{dr_process}}

\item{output_col}{\emph{string} || defaults to \code{NULL}

The name of the numerical variable from \code{process_output} that should be used to
generate the plot. This input is required for all checks \strong{EXCEPT}
\verb{Multi Site, Anomaly Detection, Longitudinal}}

\item{filter_definition}{\emph{string or vector} || defaults to \code{NULL}

A string or vector with the names of duplicate definitions that should be the focus of the analysis.
Only a single value is accepted for the following checks:
\itemize{
\item \verb{Multi Site, Anomaly Detection, Longitudinal}
\item \verb{Single Site, Anomaly Detection, Longitudinal}
}

Multiple values are accepted for:
\itemize{
\item \verb{Multi Site, Exploratory, Cross-Sectional}
\item \verb{Multi Site, Exploratory, Longitudinal}
}}

\item{text_wrapping_char}{\emph{integer} || defaults to \code{60}

An integer indicating the length limit for text wrapping on long
labels in the plots. Generally applies to the position of the
duplicate_definition on the plot, as these labels can be lengthy}

\item{large_n}{\emph{boolean} || defaults to \code{FALSE}

For Multi-Site analyses, a boolean indicating whether the large N
visualization, intended for a high volume of sites, should be used. This
visualization will produce high level summaries across all sites, with an
option to add specific site comparators via the \code{large_n_sites} parameter.}

\item{large_n_sites}{\emph{vector} || defaults to \code{NULL}

When \code{large_n = TRUE}, a vector of site names that can add site-level information
to the plot for comparison across the high level summary information.}
}
\value{
This function will produce a graph to visualize the results
from \code{dr_process} based on the parameters provided. The default
output is typically a static ggplot or gt object, but interactive
elements can be activated by passing the plot through \code{make_interactive_squba}.
For a more detailed description of output specific to each check type,
see the PEDSpace metadata repository
}
\description{
Using the tabular output generated by \code{dr_process}, this function will build a graph to
visualize the results. Each function configuration will output a bespoke ggplot. Theming can
be adjusted by the user after the graph has been output using \code{+ theme()}. Most graphs can
also be made interactive using \code{make_interactive_squba()}
}
