% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr_output_subfuncs.R
\name{dr_ss_anom_cs}
\alias{dr_ss_anom_cs}
\title{\emph{Single Site, Anomaly Detection, Cross-Sectional}}
\usage{
dr_ss_anom_cs(
  process_output,
  output_col,
  text_wrapping_char = 60L,
  facet = NULL
)
}
\arguments{
\item{process_output}{tabular output from \code{dr_process}}

\item{output_col}{name of the column from process_output to be used in the analysis.
can be any of the outlier_ or prop_outlier_ columns}

\item{text_wrapping_char}{an integer indicating the length limit for text wrapping on axis text}

\item{facet}{fields by which the graph should be facetted}
}
\value{
a bar plot displaying the number of patients, either overall or
limited to patients with at least one unmapped value, who are associated with
a number of unmapped rows that is further away from the mean than the SD
threshold
}
\description{
\emph{Single Site, Anomaly Detection, Cross-Sectional}
}
\keyword{internal}
