% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr_output_subfuncs.R
\name{dr_ss_exp_cs}
\alias{dr_ss_exp_cs}
\title{\emph{Single Site, Exploratory, Cross-Sectional}}
\usage{
dr_ss_exp_cs(
  process_output,
  output_col,
  text_wrapping_char = 60L,
  facet = NULL
)
}
\arguments{
\item{process_output}{tabular output from \code{dr_process}}

\item{output_col}{name of the column from process_output to be used in the analysis.
can be any of the numeric cols, including either proportion col or any of the
median cols}

\item{text_wrapping_char}{an integer indicating the length limit for text wrapping on axis text}

\item{facet}{fields by which the graph should be facetted}
}
\value{
a bar graph displaying either the proportion of unmapped rows/patients or the
median number of unmapped values per patient for each duplicate_definition
}
\description{
\emph{Single Site, Exploratory, Cross-Sectional}
}
\keyword{internal}
