% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr_output_subfuncs.R
\name{dr_ss_exp_la}
\alias{dr_ss_exp_la}
\title{\emph{Single Site, Exploratory, Longitudinal}}
\usage{
dr_ss_exp_la(
  process_output,
  output_col,
  text_wrapping_char = 60L,
  facet = NULL
)
}
\arguments{
\item{process_output}{tabular output from \code{dr_process}}

\item{output_col}{name of the column from process_output to be used in the analysis.
can be any of the proportion or median columns}

\item{text_wrapping_char}{an integer indicating the length limit for text wrapping on axis text}

\item{facet}{fields by which the graph should be facetted}
}
\value{
a line graph showing the proportion or median per patient of
unmapped concepts per duplicate_definition across the time series
}
\description{
\emph{Single Site, Exploratory, Longitudinal}
}
\keyword{internal}
